Revert "refactor(node-config): own the compile-time build metadata"
This reverts the prior move of vergen/build-time constants into
katana-node-config. The top-level binary should own its identity, not the
library — matching the ecosystem pattern (geth admin_nodeInfo, pathfinder's
version endpoint, reth) where the library provides a type and the binary
populates it.

Rationale: `node_getInfo` answers "what software is serving this RPC?". The
honest answer is the binary, not the library. If a downstream embedder
(Cartridge fleet, Sozo wrapper, a fork) deploys their own katana-based
binary, they want monitoring to show *their* version, not katana's. With the
library-owned design that only works if they remember to override the
BuildInfo — a silent footgun. With the binary-owned design it works by
default, and the failure mode for forgetting is "unknown" sentinels (glaring)
rather than a plausible-but-wrong katana version (insidious).

- crates/node/config: drop build.rs, vergen build-deps, const_format dep,
  BuildInfo::current(), VERSION/GIT_SHA/BUILD_TIMESTAMP consts. Struct +
  Default stay. Docstring updated to state honestly: binary populates this.
- bin/katana/build.rs: restore full vergen + DEV_BUILD_SUFFIX logic.
- bin/katana/Cargo.toml: restore [build-dependencies] with vergen/vergen-gitcl;
  drop direct katana-node-config dep (consumed via katana-cli re-export).
- bin/katana/src/cli/version.rs: restore local VERSION/GIT_SHA/BUILD_TIMESTAMP
  consts; build_info() constructs directly from them.
