+++ bin/katana/Cargo.toml
@@ -12,7 +12,6 @@ katana-chain-spec.workspace = true
 katana-cli.workspace = true
 katana-db = { workspace = true, features = [ "arbitrary" ] }
 katana-genesis.workspace = true
-katana-node-config.workspace = true
 katana-primitives.workspace = true
 katana-provider.workspace = true
 katana-rpc-types.workspace = true
@@ -48,6 +47,10 @@ url.workspace = true
 
 colored_json = { version = "5.0", optional = true }
 serde_json = { workspace = true, optional = true }
+
+[build-dependencies]
+vergen = { version = "9.0.0", features = [ "build", "cargo", "emit_and_set" ] }
+vergen-gitcl = { version = "1.0.0", features = [ "build", "cargo", "rustc", "si" ] }
 
 [dev-dependencies]
 arbitrary.workspace = true
