+++ crates/node/config/Cargo.toml
@@ -6,16 +6,11 @@ repository.workspace = true
 version.workspace = true
 
 [dependencies]
-const_format = "0.2.33"
 http.workspace = true
 serde.workspace = true
 strum.workspace = true
 strum_macros.workspace = true
 thiserror.workspace = true
-
-[build-dependencies]
-vergen = { version = "9.0.0", features = [ "build", "cargo", "emit_and_set" ] }
-vergen-gitcl = { version = "1.0.0", features = [ "build", "cargo", "rustc", "si" ] }
 
 [features]
 explorer = [  ]
